+++ localeurl/tests/test_urls.py
@@ -4,7 +4,7 @@ URLconf for testing.
 
 from django.conf.urls.defaults import *
 
-urlpatterns = patterns('localeurl.test_urls',
+urlpatterns = patterns('localeurl.tests.test_urls',
      url(r'^dummy/$', 'dummy', name='dummy0'),
      url(r'^dummy/(?P<test>.+)$', 'dummy', name='dummy1'),
 )